+++ CHANGES.rst
@@ -184,3 +184,4 @@ v1.1.6
 
 v1.1.7[unreleased]
 -----------
+* Add genre to Profile